use real assembly for the implmentation and only have some small parts in a shared project
