+++ TechTalk.SpecFlow.VSIXShared/SpecFlowPackage.cs
