+++ TechTalk.SpecFlow.VSIXShared/TechTalk.SpecFlow.VSIXShared.projitems
@@ -1,0 +1,15 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>7a62bbfa-3212-420a-ac20-63401a35f534</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>TechTalk.SpecFlow.VSIXShared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)SpecFlowPackage.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)VsContainerBuilder.cs" />
+  </ItemGroup>
+</Project>
