+++ TechTalk.SpecFlow.VsIntegration.Implementation/AssemblyAttributes.cs
@@ -1,0 +1,6 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
