+++ TechTalk.SpecFlow.VsIntegration.Implementation/DefaultDependencyProvider.cs
@@ -1,54 +1,27 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.Composition;
 using System.Linq;
 using System.Text;
+using System.Threading.Tasks;
 using BoDi;
 using EnvDTE;
 using EnvDTE80;
-using Microsoft.ApplicationInsights;
-using Microsoft.VisualStudio.Shell;
-using TechTalk.SpecFlow.IdeIntegration.Install;
 using TechTalk.SpecFlow.BindingSkeletons;
 using TechTalk.SpecFlow.IdeIntegration.Analytics;
+using TechTalk.SpecFlow.IdeIntegration.Install;
 using TechTalk.SpecFlow.IdeIntegration.Options;
 using TechTalk.SpecFlow.IdeIntegration.Tracing;
 using TechTalk.SpecFlow.VsIntegration.Analytics;
 using TechTalk.SpecFlow.VsIntegration.Install;
 using TechTalk.SpecFlow.VsIntegration.LanguageService;
 using TechTalk.SpecFlow.VsIntegration.Options;
-using TechTalk.SpecFlow.VsIntegration.TestRunner;
 using TechTalk.SpecFlow.VsIntegration.Tracing;
 using TechTalk.SpecFlow.VsIntegration.Tracing.OutputWindow;
 using TechTalk.SpecFlow.VsIntegration.Utils;
 
-namespace TechTalk.SpecFlow.VsIntegration
+namespace TechTalk.SpecFlow.VsIntegration.Implementation
 {
-    public static class VsContainerBuilder
-    {
-        internal static DefaultDependencyProvider DefaultDependencyProvider = new DefaultDependencyProvider();
-
-        public static IObjectContainer CreateContainer(SpecFlowPackagePackage package)
-        {
-            var container = new ObjectContainer();
-
-            container.RegisterInstanceAs(package);
-            container.RegisterInstanceAs<IServiceProvider>(package);
-
-            RegisterDefaults(container);
-
-            BiDiContainerProvider.CurrentContainer = container; //TODO: avoid static field
-
-            return container;
-        }
-
-        private static void RegisterDefaults(IObjectContainer container)
-        {
-            DefaultDependencyProvider.RegisterDefaults(container);
-        }
-    }
-
-    internal partial class DefaultDependencyProvider
+    public partial class DefaultDependencyProvider
     {
         static partial void RegisterCommands(IObjectContainer container);
 
@@ -94,22 +67,6 @@ namespace TechTalk.SpecFlow.VsIntegration
 
             container.RegisterInstanceAs(VsxHelper.ResolveMefDependency<IOutputWindowService>(serviceProvider));
             container.RegisterInstanceAs(VsxHelper.ResolveMefDependency<IGherkinLanguageServiceFactory>(serviceProvider));
-        }
-    }
-
-    public interface IBiDiContainerProvider
-    {
-        IObjectContainer ObjectContainer { get; }
-    }
-
-    [Export(typeof(IBiDiContainerProvider))]
-    internal class BiDiContainerProvider : IBiDiContainerProvider
-    {
-        public static IObjectContainer CurrentContainer { get; internal set; }
-
-        public IObjectContainer ObjectContainer
-        {
-            get { return CurrentContainer; }
         }
     }
 }
