+++ TechTalk.SpecFlow.VsIntegration.Implementation/LanguageService/GherkinLanguageServiceFactory.cs
@@ -6,6 +6,7 @@ using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Text;
 using TechTalk.SpecFlow.IdeIntegration.Install;
 using TechTalk.SpecFlow.IdeIntegration.Options;
+using TechTalk.SpecFlow.VsIntegration.Implementation;
 using TechTalk.SpecFlow.VsIntegration.Tracing;
 using TechTalk.SpecFlow.VsIntegration.Utils;
