+++ VsIntegration2017/TechTalk.SpecFlow.VsIntegration.2017.csproj
@@ -143,6 +143,10 @@
       <Project>{5351e241-2b8b-4d04-8999-5e3a79e8dd64}</Project>
       <Name>TechTalk.SpecFlow.VisualStudio.CodeBehindGenerator</Name>
     </ProjectReference>
+    <ProjectReference Include="..\TechTalk.SpecFlow.VsIntegration.Implementation\TechTalk.SpecFlow.VsIntegration.Implementation.csproj">
+      <Project>{10934720-f5ac-49e7-9d4f-b4a4d206c4aa}</Project>
+      <Name>TechTalk.SpecFlow.VsIntegration.Implementation</Name>
+    </ProjectReference>
   </ItemGroup>
   <PropertyGroup>
     <UseCodebase>true</UseCodebase>
@@ -154,7 +158,7 @@
       <Install>false</Install>
     </BootstrapperPackage>
   </ItemGroup>
-  <Import Project="..\TechTalk.SpecFlow.VsIntegration.Implementation\TechTalk.SpecFlow.VsIntegration.Implementation.projitems" Label="Shared" />
+  <Import Project="..\TechTalk.SpecFlow.VSIXShared\TechTalk.SpecFlow.VSIXShared.projitems" Label="Shared" />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
   <Import Project="..\TechTalk.SpecFlow.VsIntegration.Common.Build.targets" />
